Re-organização dos pacotes da tela e correção de bugs em geral

+++ src/main/java/br/unisul/revendaunisul/view/listagem/TelaListagemModelo.java
@@ -1,4 +1,4 @@
-package br.unisul.revendaunisul.view;
+package br.unisul.revendaunisul.view.listagem;
 
 import java.util.List;
 
@@ -25,6 +25,7 @@ import br.unisul.revendaunisul.entity.Modelo;
 import br.unisul.revendaunisul.enums.TipoDeVeiculo;
 import br.unisul.revendaunisul.service.MarcaService;
 import br.unisul.revendaunisul.service.ModeloService;
+import br.unisul.revendaunisul.view.cadastro.TelaCadastroModelo;
 import br.unisul.revendaunisul.view.tables.ModeloTableModel;
 
 @Component
@@ -46,13 +47,13 @@ public class TelaListagemModelo extends JFrame {
 
 	@Autowired
 	private TelaCadastroModelo cadastro;
-	
+
 	public TelaListagemModelo() {
 		setTitle("Gerenciar Modelos");
 		setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
 		setBounds(100, 100, 450, 325);
 		setLocationRelativeTo(null);
-		
+
 		contentPane = new JPanel();
 		contentPane.setBorder(new EmptyBorder(5, 5, 5, 5));
 
@@ -65,11 +66,9 @@ public class TelaListagemModelo extends JFrame {
 
 		JButton btnBuscar = new JButton("Buscar");
 		btnBuscar.addActionListener(e -> {
-			List<Modelo> modelos = service.listarPor(
-					edtFiltro.getText(),
-					(Marca) cbMarca.getSelectedItem(),
-					(TipoDeVeiculo) cbTipo.getSelectedItem());
-			
+			List<Modelo> modelos = service.listarPor(edtFiltro.getText(),
+					(Marca) cbMarca.getSelectedItem(), (TipoDeVeiculo) cbTipo.getSelectedItem());
+
 			ModeloTableModel model = new ModeloTableModel(modelos);
 			table.setModel(model);
 			TableColumnModel cm = table.getColumnModel();
@@ -89,10 +88,20 @@ public class TelaListagemModelo extends JFrame {
 			int linhaSelecionada = table.getSelectedRow();
 			ModeloTableModel model = (ModeloTableModel) table.getModel();
 			Modelo modeloSalvo = model.getBy(linhaSelecionada);
-			service.excluirPor(modeloSalvo.getId());
-			model.removeBy(linhaSelecionada);
-			table.updateUI();
-			JOptionPane.showMessageDialog(contentPane, "Modelo removido com sucesso!");
+
+			int opcaoSelecionada = JOptionPane.showConfirmDialog(contentPane,
+					"Deseja realmente remover?", "Confirmação", JOptionPane.YES_NO_OPTION);
+
+			if (opcaoSelecionada == JOptionPane.YES_OPTION) {
+				try {
+					service.excluirPor(modeloSalvo.getId());
+					model.removeBy(linhaSelecionada);
+					JOptionPane.showMessageDialog(contentPane, "Modelo removido com sucesso!");
+					table.updateUI();
+				} catch (Exception ex) {
+					JOptionPane.showMessageDialog(contentPane, ex.getMessage());
+				}
+			}
 
 		});
 
@@ -116,88 +125,81 @@ public class TelaListagemModelo extends JFrame {
 		JLabel lblTipo = new JLabel("Tipo:");
 
 		cbTipo = new JComboBox<TipoDeVeiculo>();
+		cbTipo.addItem(null);
 		for (TipoDeVeiculo t : TipoDeVeiculo.values()) {
 			cbTipo.addItem(t);
 		}
 
 		GroupLayout gl_contentPane = new GroupLayout(contentPane);
-		gl_contentPane.setHorizontalGroup(gl_contentPane.createParallelGroup(Alignment.TRAILING)
-				.addGroup(gl_contentPane.createSequentialGroup().addContainerGap()
-						.addGroup(gl_contentPane.createParallelGroup(Alignment.TRAILING)
-								.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 404,
-										Short.MAX_VALUE)
+		gl_contentPane.setHorizontalGroup(
+			gl_contentPane.createParallelGroup(Alignment.TRAILING)
+				.addGroup(gl_contentPane.createSequentialGroup()
+					.addContainerGap()
+					.addGroup(gl_contentPane.createParallelGroup(Alignment.LEADING)
+						.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 400, Short.MAX_VALUE)
+						.addGroup(Alignment.TRAILING, gl_contentPane.createSequentialGroup()
+							.addComponent(btnEditar)
+							.addPreferredGap(ComponentPlacement.RELATED)
+							.addComponent(btnExcluir)
+							.addPreferredGap(ComponentPlacement.RELATED)
+							.addComponent(btnNovo, GroupLayout.PREFERRED_SIZE, 65, GroupLayout.PREFERRED_SIZE))
+						.addGroup(gl_contentPane.createSequentialGroup()
+							.addGroup(gl_contentPane.createParallelGroup(Alignment.TRAILING)
+								.addComponent(LblNome)
+								.addComponent(LblMarca))
+							.addPreferredGap(ComponentPlacement.RELATED)
+							.addGroup(gl_contentPane.createParallelGroup(Alignment.LEADING)
 								.addGroup(gl_contentPane.createSequentialGroup()
-										.addComponent(btnEditar, GroupLayout.PREFERRED_SIZE, 65,
-												GroupLayout.PREFERRED_SIZE)
-										.addPreferredGap(ComponentPlacement.RELATED)
-										.addComponent(btnExcluir, GroupLayout.PREFERRED_SIZE, 65,
-												GroupLayout.PREFERRED_SIZE)
-										.addPreferredGap(ComponentPlacement.RELATED)
-										.addComponent(btnNovo, GroupLayout.PREFERRED_SIZE, 65,
-												GroupLayout.PREFERRED_SIZE))
-								.addGroup(Alignment.LEADING, gl_contentPane.createSequentialGroup()
-										.addGroup(gl_contentPane
-												.createParallelGroup(Alignment.LEADING)
-												.addComponent(LblNome).addComponent(LblMarca,
-														GroupLayout.PREFERRED_SIZE, 34,
-														GroupLayout.PREFERRED_SIZE))
-										.addPreferredGap(ComponentPlacement.RELATED)
-										.addGroup(gl_contentPane
-												.createParallelGroup(Alignment.LEADING)
-												.addGroup(gl_contentPane.createSequentialGroup()
-														.addComponent(cbMarca,
-																GroupLayout.PREFERRED_SIZE, 130,
-																GroupLayout.PREFERRED_SIZE)
-														.addPreferredGap(ComponentPlacement.RELATED)
-														.addComponent(lblTipo,
-																GroupLayout.PREFERRED_SIZE, 28,
-																GroupLayout.PREFERRED_SIZE)
-														.addPreferredGap(ComponentPlacement.RELATED)
-														.addComponent(cbTipo, 0, 129,
-																Short.MAX_VALUE))
-												.addComponent(edtFiltro, Alignment.TRAILING,
-														GroupLayout.DEFAULT_SIZE, 295,
-														Short.MAX_VALUE))
-										.addPreferredGap(ComponentPlacement.RELATED)
-										.addComponent(btnBuscar)))
-						.addContainerGap()));
-		gl_contentPane.setVerticalGroup(gl_contentPane.createParallelGroup(Alignment.LEADING)
-				.addGroup(gl_contentPane.createSequentialGroup().addContainerGap()
-						.addGroup(gl_contentPane.createParallelGroup(Alignment.BASELINE)
-								.addComponent(edtFiltro, GroupLayout.PREFERRED_SIZE,
-										GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
-								.addComponent(LblNome).addComponent(btnBuscar))
-						.addPreferredGap(ComponentPlacement.RELATED)
-						.addGroup(gl_contentPane.createParallelGroup(Alignment.BASELINE)
-								.addComponent(LblMarca)
-								.addComponent(cbMarca, GroupLayout.PREFERRED_SIZE,
-										GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
-								.addComponent(cbTipo, GroupLayout.PREFERRED_SIZE,
-										GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
-								.addComponent(lblTipo))
-						.addGap(14)
-						.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 160, Short.MAX_VALUE)
-						.addPreferredGap(ComponentPlacement.RELATED)
-						.addGroup(gl_contentPane.createParallelGroup(Alignment.BASELINE)
-								.addComponent(btnNovo).addComponent(btnExcluir)
-								.addComponent(btnEditar))
-						.addContainerGap()));
+									.addComponent(cbMarca, GroupLayout.PREFERRED_SIZE, 136, GroupLayout.PREFERRED_SIZE)
+									.addPreferredGap(ComponentPlacement.RELATED)
+									.addComponent(lblTipo, GroupLayout.PREFERRED_SIZE, 28, GroupLayout.PREFERRED_SIZE)
+									.addGap(6)
+									.addComponent(cbTipo, 0, 86, Short.MAX_VALUE))
+								.addComponent(edtFiltro, GroupLayout.DEFAULT_SIZE, 268, Short.MAX_VALUE))
+							.addPreferredGap(ComponentPlacement.RELATED)
+							.addComponent(btnBuscar, GroupLayout.PREFERRED_SIZE, 75, GroupLayout.PREFERRED_SIZE)))
+					.addContainerGap())
+		);
+		gl_contentPane.setVerticalGroup(
+			gl_contentPane.createParallelGroup(Alignment.LEADING)
+				.addGroup(gl_contentPane.createSequentialGroup()
+					.addContainerGap()
+					.addGroup(gl_contentPane.createParallelGroup(Alignment.BASELINE)
+						.addComponent(btnBuscar)
+						.addComponent(LblNome)
+						.addComponent(edtFiltro, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE))
+					.addPreferredGap(ComponentPlacement.RELATED)
+					.addGroup(gl_contentPane.createParallelGroup(Alignment.BASELINE)
+						.addComponent(cbTipo, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
+						.addComponent(cbMarca, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
+						.addComponent(lblTipo)
+						.addComponent(LblMarca))
+					.addGap(14)
+					.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 149, Short.MAX_VALUE)
+					.addPreferredGap(ComponentPlacement.RELATED)
+					.addGroup(gl_contentPane.createParallelGroup(Alignment.BASELINE)
+						.addComponent(btnNovo)
+						.addComponent(btnExcluir)
+						.addComponent(btnEditar))
+					.addContainerGap())
+		);
 		contentPane.setLayout(gl_contentPane);
 	}
-	
+
 	private void carregarOpcoes() {
 		marcas = marcaService.listarTodos();
 		cbMarca.removeAllItems();
+		cbMarca.addItem(null);
 		for (Marca m : marcas) {
 			cbMarca.addItem(m);
 		}
 	}
-	
+
 	@Override
 	public void setVisible(boolean b) {
 		this.carregarOpcoes();
 		this.cbMarca.setSelectedItem(0);
 		super.setVisible(b);
 	}
-	
+
 }
