+++ src/main/java/br/unisul/revendaunisul/view/listagem/TelaListagemVeiculo.java
@@ -1,4 +1,4 @@
-package br.unisul.revendaunisul.view;
+package br.unisul.revendaunisul.view.listagem;
 
 import java.awt.event.ItemEvent;
 import java.util.List;
@@ -26,6 +26,7 @@ import br.unisul.revendaunisul.entity.Veiculo;
 import br.unisul.revendaunisul.service.MarcaService;
 import br.unisul.revendaunisul.service.ModeloService;
 import br.unisul.revendaunisul.service.VeiculoService;
+import br.unisul.revendaunisul.view.cadastro.TelaCadastroVeiculo;
 import br.unisul.revendaunisul.view.tables.VeiculoTableModel;
 
 @Component
@@ -41,7 +42,7 @@ public class TelaListagemVeiculo extends JFrame {
 
 	@Autowired
 	private VeiculoService service;
-	
+
 	@Autowired
 	private ModeloService modeloService;
 
@@ -50,7 +51,7 @@ public class TelaListagemVeiculo extends JFrame {
 
 	@Autowired
 	private TelaCadastroVeiculo cadastro;
-	
+
 	public TelaListagemVeiculo() {
 		setTitle("Gerenciar Veículos");
 		setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
@@ -63,9 +64,8 @@ public class TelaListagemVeiculo extends JFrame {
 
 		JButton btnBuscar = new JButton("Buscar");
 		btnBuscar.addActionListener(e -> {
-			List<Veiculo> veiculos = service.listarPor(
-					(Modelo) cbModelo.getSelectedItem());
-			
+			List<Veiculo> veiculos = service.listarPor((Modelo) cbModelo.getSelectedItem());
+
 			VeiculoTableModel model = new VeiculoTableModel(veiculos);
 			table.setModel(model);
 			TableColumnModel cm = table.getColumnModel();
@@ -81,15 +81,24 @@ public class TelaListagemVeiculo extends JFrame {
 		});
 
 		JButton btnExcluir = new JButton("Excluir");
-		btnExcluir.addActionListener(e -> {
+		btnExcluir.addActionListener(event -> {
 			int linhaSelecionada = table.getSelectedRow();
 			VeiculoTableModel model = (VeiculoTableModel) table.getModel();
 			Veiculo veiculoSalvo = model.getBy(linhaSelecionada);
-			service.excluirPor(veiculoSalvo.getId());
-			model.removeBy(linhaSelecionada);
-			table.updateUI();
-			JOptionPane.showMessageDialog(contentPane, "Veículo removido com sucesso!");
 
+			int opcaoSelecionada = JOptionPane.showConfirmDialog(contentPane,
+					"Deseja realmente remover?", "Confirmação", JOptionPane.YES_NO_OPTION);
+
+			if (opcaoSelecionada == JOptionPane.YES_OPTION) {
+				try {
+					service.excluirPor(veiculoSalvo.getId());
+					model.removeBy(linhaSelecionada);
+					table.updateUI();
+					JOptionPane.showMessageDialog(contentPane, "Veículo removido com sucesso!");
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
+			}
 		});
 
 		JButton btnEditar = new JButton("Editar");
@@ -106,19 +115,21 @@ public class TelaListagemVeiculo extends JFrame {
 		table.setFillsViewportHeight(true);
 
 		JLabel lblMarca = new JLabel("Modelo:");
-		
+
 		cbModelo = new JComboBox<Modelo>();
 
 		JLabel LblMarca = new JLabel("Marca:");
 
 		cbMarca = new JComboBox<Marca>();
-		
+
 		cbMarca.addItemListener(e -> {
 			if (e.getStateChange() == ItemEvent.SELECTED) {
-				cbModelo.removeAllItems();
-				modelos = modeloService.listarPor((Marca) cbMarca.getSelectedItem());
-				for (Modelo m : modelos) {
-					cbModelo.addItem(m);
+				if (!cbMarca.getSelectedItem().equals(0)) {
+					cbModelo.removeAllItems();
+					modelos = modeloService.listarPor((Marca) cbMarca.getSelectedItem());
+					for (Modelo m : modelos) {
+						cbModelo.addItem(m);
+					}
 				}
 			}
 		});
@@ -128,24 +139,24 @@ public class TelaListagemVeiculo extends JFrame {
 			gl_contentPane.createParallelGroup(Alignment.TRAILING)
 				.addGroup(gl_contentPane.createSequentialGroup()
 					.addContainerGap()
-					.addGroup(gl_contentPane.createParallelGroup(Alignment.LEADING)
-						.addGroup(Alignment.TRAILING, gl_contentPane.createSequentialGroup()
-							.addComponent(btnEditar, GroupLayout.PREFERRED_SIZE, 65, GroupLayout.PREFERRED_SIZE)
+					.addGroup(gl_contentPane.createParallelGroup(Alignment.TRAILING)
+						.addGroup(gl_contentPane.createSequentialGroup()
+							.addComponent(btnEditar)
 							.addPreferredGap(ComponentPlacement.RELATED)
-							.addComponent(btnExcluir, GroupLayout.PREFERRED_SIZE, 65, GroupLayout.PREFERRED_SIZE)
+							.addComponent(btnExcluir)
 							.addPreferredGap(ComponentPlacement.RELATED)
 							.addComponent(btnNovo, GroupLayout.PREFERRED_SIZE, 65, GroupLayout.PREFERRED_SIZE))
-						.addGroup(Alignment.TRAILING, gl_contentPane.createSequentialGroup()
-							.addComponent(LblMarca, GroupLayout.PREFERRED_SIZE, 34, GroupLayout.PREFERRED_SIZE)
-							.addPreferredGap(ComponentPlacement.RELATED)
+						.addGroup(gl_contentPane.createSequentialGroup()
+							.addComponent(LblMarca)
+							.addGap(7)
 							.addComponent(cbMarca, GroupLayout.PREFERRED_SIZE, 96, GroupLayout.PREFERRED_SIZE)
-							.addPreferredGap(ComponentPlacement.RELATED)
+							.addGap(6)
 							.addComponent(lblMarca)
-							.addPreferredGap(ComponentPlacement.RELATED)
-							.addComponent(cbModelo, 0, 153, Short.MAX_VALUE)
+							.addGap(6)
+							.addComponent(cbModelo, 0, 121, Short.MAX_VALUE)
 							.addPreferredGap(ComponentPlacement.RELATED)
 							.addComponent(btnBuscar))
-						.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 404, Short.MAX_VALUE))
+						.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 400, Short.MAX_VALUE))
 					.addContainerGap())
 		);
 		gl_contentPane.setVerticalGroup(
@@ -155,31 +166,30 @@ public class TelaListagemVeiculo extends JFrame {
 					.addGroup(gl_contentPane.createParallelGroup(Alignment.TRAILING)
 						.addGroup(gl_contentPane.createSequentialGroup()
 							.addGroup(gl_contentPane.createParallelGroup(Alignment.BASELINE)
-								.addComponent(LblMarca)
 								.addComponent(cbMarca, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
-								.addComponent(lblMarca))
-							.addPreferredGap(ComponentPlacement.RELATED, 3, GroupLayout.PREFERRED_SIZE))
-						.addGroup(gl_contentPane.createParallelGroup(Alignment.BASELINE)
-							.addComponent(btnBuscar)
-							.addComponent(cbModelo, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)))
+								.addComponent(LblMarca)
+								.addComponent(lblMarca)
+								.addComponent(cbModelo, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE))
+							.addPreferredGap(ComponentPlacement.RELATED))
+						.addComponent(btnBuscar))
 					.addGap(18)
-					.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 160, Short.MAX_VALUE)
+					.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 170, Short.MAX_VALUE)
 					.addPreferredGap(ComponentPlacement.UNRELATED)
 					.addGroup(gl_contentPane.createParallelGroup(Alignment.BASELINE)
 						.addComponent(btnNovo)
-						.addComponent(btnExcluir)
-						.addComponent(btnEditar))
+						.addComponent(btnEditar)
+						.addComponent(btnExcluir))
 					.addContainerGap())
 		);
 		contentPane.setLayout(gl_contentPane);
 	}
-	
+
 	@Override
 	public void setVisible(boolean b) {
 		this.carregarOpcoes();
 		super.setVisible(b);
 	}
-	
+
 	private void carregarOpcoes() {
 		marcas = marcaService.listarTodos();
 		cbMarca.removeAllItems();
@@ -187,5 +197,5 @@ public class TelaListagemVeiculo extends JFrame {
 			cbMarca.addItem(m);
 		}
 	}
-	
+
 }
